feat: moved the input for creating todo into a separate component
