+++ src/components/TodoList/styled.ts
@@ -59,44 +59,6 @@ export const TodoListDayDescription = styled.p`
   font-weight: 600;
 `;
 
-export const TodoListWrapper = styled.div`
-  ${flex('space-between', 'stretch')}
-
-  gap: 10px;
-  margin-bottom: 10px;
-`;
-
-export const TodoListInput = styled.input`
-  width: 100%;
-
-  padding: 8px 10px;
-
-  border-radius: ${borderRadius.high};
-  border: 1px solid ${colors.greyLight};
-  outline: none;
-`;
-
-export const TodoListAddTodoButton = styled.button`
-  padding: 8px 10px;
-
-  color: ${colors.white};
-  font-weight: 600;
-
-  background-color: ${colors.blue};
-  border: none;
-  border-radius: ${borderRadius.high};
-  transition: opacity 200ms;
-  cursor: pointer;
-
-  &:hover {
-    opacity: ${opacity.low};
-  }
-
-  &:active {
-    opacity: ${opacity.high};
-  }
-`;
-
 export const TodoListOwn = styled.ul`
   ${flex()}
